ref: inline theme.tsx into theme.chonk.tsx
